refactor: 로또 객체 내에서 번호 발급
